pyproject.toml
[tool.poetry]
name = "telegram-bot-boilerplate"
version = "0.1.0"
description = "A simple boilerplate for building Telegram bots"
authors = ["2rage <the2rage@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.9"
pydantic = "^2.10.3"
pydantic-settings = "^2.6.1"
watchfiles = "^1.0.0"
asyncpg = "^0.30.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
